backend-python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
groq>=0.9.0
requests==2.31.0
gunicorn==21.2.0
Pillow>=10.0.0
watchdog>=4.0.0
